refactor: reorganize BackgroundJobsSample structure
- Separate requests into Requests/ folder (SendEmailRequest, ProcessPaymentRequest)
- Separate notifications into Notifications/ folder (OrderCreatedNotification)
- Separate handlers into Handlers/ folder (all handler implementations)
- Keep Program.cs clean with only configuration and execution logic
- Follow same organizational pattern as SimpleSample and AdvancedSample
- Improve code maintainability and readability
