+++ samples/BackgroundJobsSample/Program.cs
@@ -1,3 +1,6 @@
+using BackgroundJobsSample.Handlers;
+using BackgroundJobsSample.Notifications;
+using BackgroundJobsSample.Requests;
 using Coordix.Background.Extensions;
 using Coordix.Background.Interfaces;
 using Coordix.Extensions;
@@ -8,118 +11,6 @@ using Microsoft.Extensions.Logging;
 
 namespace BackgroundJobsSample;
 
-// Example requests and notifications
-public class SendEmailRequest : IRequest
-{
-	public string To { get; set; } = string.Empty;
-	public string Subject { get; set; } = string.Empty;
-	public string Body { get; set; } = string.Empty;
-}
-
-public class ProcessPaymentRequest : IRequest<string>
-{
-	public decimal Amount { get; set; }
-	public string CustomerId { get; set; } = string.Empty;
-}
-
-public class OrderCreatedNotification : INotification
-{
-	public string OrderId { get; set; } = string.Empty;
-	public string CustomerId { get; set; } = string.Empty;
-	public decimal Total { get; set; }
-}
-
-// Handlers
-public class SendEmailHandler : IRequestHandler<SendEmailRequest>
-{
-	private readonly ILogger<SendEmailHandler> _logger;
-
-	public SendEmailHandler(ILogger<SendEmailHandler> logger)
-	{
-		_logger = logger;
-	}
-
-	public Task Handle(SendEmailRequest request, CancellationToken cancellationToken)
-	{
-		_logger.LogInformation("📧 Sending email to {To} with subject: {Subject}", request.To, request.Subject);
-		_logger.LogInformation("   Body: {Body}", request.Body);
-
-		// Simulate email sending
-		Task.Delay(500, cancellationToken).Wait(cancellationToken);
-
-		_logger.LogInformation("✅ Email sent successfully to {To}", request.To);
-		return Task.CompletedTask;
-	}
-}
-
-public class ProcessPaymentHandler : IRequestHandler<ProcessPaymentRequest, string>
-{
-	private readonly ILogger<ProcessPaymentHandler> _logger;
-
-	public ProcessPaymentHandler(ILogger<ProcessPaymentHandler> logger)
-	{
-		_logger = logger;
-	}
-
-	public Task<string> Handle(ProcessPaymentRequest request, CancellationToken cancellationToken)
-	{
-		_logger.LogInformation("💳 Processing payment of {Amount} for customer {CustomerId}",
-			request.Amount, request.CustomerId);
-
-		// Simulate payment processing
-		Task.Delay(1000, cancellationToken).Wait(cancellationToken);
-
-		var transactionId = $"TXN-{Guid.NewGuid():N}";
-		_logger.LogInformation("✅ Payment processed successfully. Transaction ID: {TransactionId}", transactionId);
-
-		return Task.FromResult(transactionId);
-	}
-}
-
-public class OrderCreatedHandler : INotificationHandler<OrderCreatedNotification>
-{
-	private readonly ILogger<OrderCreatedHandler> _logger;
-
-	public OrderCreatedHandler(ILogger<OrderCreatedHandler> logger)
-	{
-		_logger = logger;
-	}
-
-	public Task Handle(OrderCreatedNotification notification, CancellationToken cancellationToken)
-	{
-		_logger.LogInformation("📦 Order {OrderId} created for customer {CustomerId} with total {Total}",
-			notification.OrderId, notification.CustomerId, notification.Total);
-
-		// Simulate order processing
-		Task.Delay(300, cancellationToken).Wait(cancellationToken);
-
-		_logger.LogInformation("✅ Order {OrderId} processed successfully", notification.OrderId);
-		return Task.CompletedTask;
-	}
-}
-
-public class OrderCreatedInventoryHandler : INotificationHandler<OrderCreatedNotification>
-{
-	private readonly ILogger<OrderCreatedInventoryHandler> _logger;
-
-	public OrderCreatedInventoryHandler(ILogger<OrderCreatedInventoryHandler> logger)
-	{
-		_logger = logger;
-	}
-
-	public Task Handle(OrderCreatedNotification notification, CancellationToken cancellationToken)
-	{
-		_logger.LogInformation("📊 Updating inventory for order {OrderId}", notification.OrderId);
-
-		// Simulate inventory update
-		Task.Delay(200, cancellationToken).Wait(cancellationToken);
-
-		_logger.LogInformation("✅ Inventory updated for order {OrderId}", notification.OrderId);
-		return Task.CompletedTask;
-	}
-}
-
-// Main program
 public class Program
 {
 	public static async Task Main(string[] args)
@@ -195,4 +86,3 @@ public class Program
 		await host.StopAsync();
 	}
 }
-
