+++ samples/BackgroundJobsSample/Requests/SendEmailRequest.cs
@@ -1,0 +1,11 @@
+using Coordix.Interfaces;
+
+namespace BackgroundJobsSample.Requests;
+
+public sealed class SendEmailRequest : IRequest
+{
+	public string To { get; set; } = string.Empty;
+	public string Subject { get; set; } = string.Empty;
+	public string Body { get; set; } = string.Empty;
+}
+
